First Release
tested on a MacBook Air with a canable 2.0 device in canfd mode.  supports message highlighting on change, graphing of bytes, sending of messages and interval sending.
Engineered with the assistance of GROK!  This is probably the best free can tool for Mac. please make it even better!  performance seems quite good.